Fix two -Werror failures, and type-check the RDMA headers locally
Both errors were mine this time. The cache-half diagnostics are gone, so
the SYSTEM include change did its job.

ds_rdma.hpp handed ibv_wc_status straight to std::to_string, which picks
the int overload and so trips -Werror=sign-promo. Widened explicitly.
main.cpp passed layout.cache_layers (uint64_t) to headAddrs (uint32_t);
that one was only a warning, since -Wno-error=conversion is in the flag
set, but it is a real narrowing and is now an explicit cast.

The larger problem is that this is the third cluster build spent on
errors a compiler catches in a second, and ds_rdma.hpp cannot be compiled
off-cluster at all -- dory and libibverbs are not here. So
disco-skip/tests/dory_stub/ now declares just enough of
dory::conn::ReliableConnection and ibverbs for those headers to compile,
and `make rdma` builds ds_rdma.hpp, the bootstrap write loop and the
verifier-over-RDMA combination against it under the strict warning set.

Compile-only, and deliberately not a fake fabric: nothing in the stub
moves data and no test asserts behaviour through it.

I checked whether it would have caught today's failures, and it would
have caught neither -- so the README says so rather than leaving an
overstated claim in the repo. sign-promo is GCC-only and there is no GCC
on the dev Mac; the narrowing is in main.cpp, which also needs lyra, fmt
and the dory control plane and is past what is worth stubbing. What the
stub does buy is type-checking of the RDMA-facing headers: argument
order, signatures drifting from conn/src/rc.hpp, templates that do not
instantiate, and narrowing of runtime values. That pays off mostly in
Step 2, where the descent state machine is large and entirely
dory-facing.

Same correction applies to kNodeRecordBytes, added here to name the
sizeof -> uint32_t bound for ibverbs sge lengths: I first described it as
a fix, but sizeof is a constant that provably fits and no compiler warns
about it. It documents the bound and nothing more, and the comment now
says that.

Verified: make -C disco-skip/tests (toggle combinations, verifier suite,
strict headers, RDMA stub compile) and make -C disco-skip/include check
all pass.

+++ disco-skip/src/main.cpp
@@ -541,7 +541,8 @@ int main(int argc, char** argv) {
         // that was never told them reports a miss for every key left of the
         // first boundary at every level, which is correct but silently slow
         // (interface doc §5).
-        ds::bootstrapHeads(state.cache_sv, ds::headAddrs(layout.cache_layers));
+        ds::bootstrapHeads(state.cache_sv,
+                           ds::headAddrs(static_cast<uint32_t>(layout.cache_layers)));
         if (!ds::headsBootstrapped(state.cache_sv)) {
             std::cerr << "Cache heads were not bootstrapped" << std::endl;
             return 1;
